requirements.txt
gunicorn==20.0.4
django==3.0.5
dj-database-url==0.5.0
celery==4.4.2
psycopg2==2.8.4
whitenoise==5.0.1
python-dotenv==0.12.0